% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defectAssessment.R
\name{defectAssessment}
\alias{defectAssessment}
\title{Defect assessment and summary}
\usage{
defectAssessment(seasons, password)
}
\arguments{
\item{seasons}{a vector of the seasons to be returned.}

\item{password}{a text string containing the password to the SQL database (see administrator for the password string)}
}
\value{
a tibble with each row representing a specific defect for a specific grader batch.
}
\description{
`defectAssessment` returns a tibble which contains the summary information on all defects
}
\details{
The `defectAssessment` function is a helper function to return a useful tibble to allow more detailed analysis.
The function returns all recorded defects by season and grader batch.  It also integrates the pack-out value to
deliver the defect proportion of the entire batch.
}
\seealso{
[graderFunction()], [binsHarvested()], [growerRTE()]
}
