% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/graderFunction.R
\name{graderFunction}
\alias{graderFunction}
\title{Grader batch summary function}
\usage{
graderFunction(seasons, password)
}
\arguments{
\item{seasons}{a vector of the seasons to be returned.}

\item{password}{a text string containing the password to the SQL database (see administrator for the password string)}
}
\value{
a tibble with each row representing a specific bin consignment.
}
\description{
`graderFunction` returns a tibble which contains the summary information on all batches processed through the packhouse
}
\details{
The `graderFunction` function is a helper function to return a useful tibble to allow more detailed analysis
of bins processed through the packhouse and ABC system. The function returns all recorded batches by season
and grader batch number.
}
\seealso{
[binsHarvested()], [defectAssessment()], [growerRTE()]
}
