% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/growerRTE.R
\name{growerRTE}
\alias{growerRTE}
\title{Grower RTE summary function}
\usage{
growerRTE(seasons, password)
}
\arguments{
\item{seasons}{a vector of the seasons to be returned.}

\item{password}{the password string to the SQL database (see administrator for the password string)}
}
\value{
a tibble with each row representing a specific grader batch.
}
\description{
`growerRTEs` returns a tibble which contains the summary information on grower RTEs including the growerRTE category
}
\details{
The `growerRTE` function is a helper function to return a useful tibble to allow more detailed analysis
of growerRTEs for the various pack categories (i.e. tubes packed, re-packs, export bins and pre-sized bins). The function returns all grower RTEs by GraderBatchID and orchard and block information.
}
\seealso{
[graderFunction()], [defectAssessment()], [binsHarvested()]
}
